+++ apps/tergite-mss/start_mss.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+MSS_CONFIG_FILE=$([[ -z "$MSS_CONFIG_FILE" ]] && echo "mss-config.toml" || echo "$MSS_CONFIG_FILE")
+
 exit_with_error () {
   echo "$1"
   exit 1
@@ -7,8 +9,8 @@ exit_with_error () {
 
 extract_env_var () {
   local env_name="$1"
-  local res=$(grep "^[[:space:]]*${env_name}=" .env | grep -v '^[[:space:]]*#' | sed "s/^[[:space:]]*${env_name}=//" | head -n 1)
-  [[ -z "$res" ]]  &&  exit_with_error "Config Error: Use ${env_name}=<value> in the .env file."
+  local res=$(grep "^[[:space:]]*${env_name}[[:space:]]*=" "$MSS_CONFIG_FILE" | grep -v '^[[:space:]]*#' | sed "s/^[[:space:]]*${env_name}[[:space:]]*=[[:space:]]*//" | head -n 1)
+  [[ -z "$res" ]]  &&  exit_with_error "Config Error: Use ${env_name}=<value> in the $MSS_CONFIG_FILE file."
   echo $res
 }
 
@@ -41,9 +43,12 @@ wait_for_process()
 }
 
 # port handling
-PORT_NUMBER=$([[ "$MSS_PORT" ]] && echo $MSS_PORT || echo $(extract_env_var "MSS_PORT"))
-[[ ! "$PORT_NUMBER" =~ ^[0-9]+$ ]]  &&  exit_with_error "Port configuration failed. Use MSS_PORT=<int> in the .env file."
+PORT_NUMBER=$([[ "$MSS_PORT" ]] && echo $MSS_PORT || echo $(extract_env_var "mss_port"))
+[[ ! "$PORT_NUMBER" =~ ^[0-9]+$ ]]  &&  exit_with_error "Port configuration failed. Use mss_port=<int> in the $MSS_CONFIG_FILE file."
 
+# app settings
+APP_SETTINGS=$(echo $(extract_env_var "environment"))
+[[ -z "$APP_SETTINGS" ]]  &&  exit_with_error "Port configuration failed. Use environment=<string> in the $MSS_CONFIG_FILE file."
 
 set_sigterm_handler