+++ apps/tergite-webgui/cypress/e2e/linechart.cy.ts
@@ -11,7 +11,7 @@ meResponses.forEach((resp) => {
 	describe(`Line chart screen for user ${id}`, () => {
 		beforeEach(() => {
 			const baseUrl = Cypress.env('API_BASE_URL');
-			const oauthConfigFile = Cypress.env('AUTH_CONFIG_FILE');
+			const mssConfigFile = Cypress.env('MSS_CONFIG_FILE');
 			const landingPageUrl = Cypress.env('LANDING_ENDPOINT');
 			const apiRoutes = getApiRoutes(baseUrl);
 
@@ -34,11 +34,11 @@ meResponses.forEach((resp) => {
 			}).as('luki-type4-codomain-request');
 
 			if (user.id) {
-				cy.task('readToml', oauthConfigFile).then((oauthConfig: Record<string, any>) => {
-					cy.wrap(generateJwt(user, oauthConfig)).then((jwtToken: string) => {
-						const generalConfig = oauthConfig.general || {};
-						const cookieName = generalConfig.cookie_name;
-						const domain = generalConfig.cookie_domain;
+				cy.task('readToml', mssConfigFile).then((mssConfig: Record<string, any>) => {
+					cy.wrap(generateJwt(user, mssConfig)).then((jwtToken: string) => {
+						const authConfig = mssConfig.auth || {};
+						const cookieName = authConfig.cookie_name;
+						const domain = authConfig.cookie_domain;
 
 						cy.setCookie(cookieName, jwtToken, {
 							domain,