+++ apps/tergite-webgui/cypress/support/utils.ts
@@ -5,13 +5,13 @@ type _User = { id: string; roles: string[] };
 /**
  * Generate a valid test JWT for the given user
  * @param user - the user for whom the JWT is generated
- * @param oauthConfig - the auth config got from the auth config file
+ * @param mssConfig - the MSS config got from the mss-config file
  * @returns the JSON web token
  */
-export async function generateJwt(user: _User, oauthConfig: Record<string, any>) {
+export async function generateJwt(user: _User, mssConfig: Record<string, any>) {
 	const payload = { sub: user.id, roles: [...user.roles] };
-	const generalConfig = oauthConfig.general || {};
-	const jwtSecret = generalConfig.jwt_secret || '';
+	const authConfig = mssConfig.auth || {};
+	const jwtSecret = authConfig.jwt_secret || '';
 	const secret = new TextEncoder().encode(jwtSecret);
 
 	const alg = 'HS256';
@@ -24,3 +24,4 @@ export async function generateJwt(user: _User, oauthConfig: Record<string, any>)
 		.setExpirationTime('1h')
 		.sign(secret);
 }
+