+++ apps/tergite-webgui/src/pages/api/logout/index.page.ts
@@ -9,11 +9,11 @@ export default async function handler(
 		res.status(405).json({ detail: 'METHOD NOT ALLOWED' });
 		return;
 	}
-	const oauthConfigFile = process.env.AUTH_CONFIG_FILE || 'auth_config.toml';
-	const oauthConfig = await readToml(oauthConfigFile);
+	const mssConfigFile = process.env.MSS_CONFIG_FILE || 'mss-config.toml';
+	const mssConfig = await readToml(mssConfigFile);
 
-	const generalConfig = oauthConfig.general || {};
-	const cookieName = generalConfig.cookie_name;
+	const authConfig = mssConfig.auth || {};
+	const cookieName = authConfig.cookie_name;
 	logout(req, res, cookieName);
 
 	res.status(200).json({ message: 'LOGGED OUT' });